Update API AddImage: update request parameters body.Size' format has changed.

+++ aiworkspace-20210204/core/Models/AddImageRequest.cs
@@ -124,7 +124,7 @@ namespace AlibabaCloud.SDK.AIWorkSpace20210204.Models
         /// </summary>
         [NameInMap("Size")]
         [Validation(Required=false)]
-        public int? Size { get; set; }
+        public long? Size { get; set; }
 
         [NameInMap("SourceId")]
         [Validation(Required=false)]
